API-2497: Hooks Migration Navbar & HoverImage Components (#473)
* Move NavBar component to navBar folder

* Migrate NavBar to hooks/func component

* Move HoverImage to hoverImage folder

* Migrate HoverImage component from class to hooks
