+++ src/components/index.ts
@@ -1,7 +1,9 @@
-export * from './header/Header';
 export * from './authorizationCard/AuthorizationCard';
-export * from './pageContent/PageContent';
 export * from './groupedAccordions/GroupedAccordions';
+export * from './header/Header';
+export * from './hoverImage/HoverImage';
+export * from './navBar/NavBar';
+export * from './pageContent/PageContent';
 export * from './search/Search';
 export * from './form/Form';
 export * from './onlyTags/OnlyTags';
