+++ src/containers/News.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
 import * as React from 'react';
 
+import { HoverImage } from '../components';
 import CardLink from '../components/CardLink';
 import EmbeddedYoutubeVideo from '../components/EmbeddedYoutubeVideo';
-import HoverImage from '../components/HoverImage';
 import PageHeader from '../components/PageHeader';
 import SideNav, { SideNavEntry } from '../components/SideNav';
 import * as NewsData from '../content/news.yml';
@@ -64,7 +64,6 @@ const MediaItem = ({ item }: { item: NewsItem }): JSX.Element => {
 };
 
 const ItemDescription = ({ item }: { item: NewsItem }): JSX.Element => (
-
   <p>
     <a href={item.url}>{item.title}</a>
     <br />
